requirements.txt
fastapi~=0.115.4
python-dotenv~=1.0.1
lancedb~=0.13.0
langchain-openai~=0.2.6
tqdm~=4.67.0
pandas~=2.2.3
pyarrow~=15.0.2
numpy~=1.26.4
requests~=2.32.3
docling~=2.4.2
pydantic~=2.9.2
graphrag~=0.4.1
transformers~=4.46.2
